+++ ros_bt_py/ros_bt_py/ros_nodes/subtree.py
@@ -107,9 +107,11 @@ class Subtree(Leaf):
             logging_manager=self.subtree_logging_manager,
         )
 
-        load_result = self.load_subtree()
-        if load_result.is_err():
-            raise load_result.unwrap_err()
+        match self.load_subtree():
+            case Err(e):
+                raise e
+            case Ok(None):
+                pass
 
         if self.subtree_manager:
             self.subtree_manager.add_subtree_structure(